2 reorginized weapons_repository in two separated abc_product_repository and abc_property_repository
