+++ src/repositories/weapons_repository.py
@@ -1,74 +1,19 @@
-from models.product import Product
-from sqlalchemy.orm import Session, joinedload
-from infrastructures.product_exception import *
-from infrastructures.weapons_exception import weaponsException404
+from abstracts.abc_product_repository import AbcProductRepository
+from abstracts.abc_property_repository import AbcPropertyRepository
 from abstracts.abc_weapons_repository import AbcWeaponsRepository
-from models.property import Property
-from schemas.product_schema import ProductSchema, ProductSchemaModify
+from sqlalchemy.orm import Session
+from repositories.product_repository import ProductRepository
+from repositories.property_repository import PropertyRepository
 
 class SqlDbWeaponsRepository(AbcWeaponsRepository):
 
     def __init__(self, db: Session):
-        self.db=db
+        self.db = db
 
-    def get_all(self):
-        products = self.db.query(Product).all()
-        if not products:
-            raise products_empty_exc
+    @property
+    def product_repository(self) ->AbcProductRepository:
+        return ProductRepository(self.db)
 
-        return products
-
-    def get_product_card_by_name(self, model: str):
-        product = self.db.query(Product).options(joinedload(Product.property)).filter(Product.model == model).first()
-        if not product:
-            raise weaponsException404(name_obj='product', _property=model)
-
-        return product
-
-    def delete_product_by_name(self, model: str):
-        product = self.db.query(Product).filter(Product.model == model).first()
-        if not product:
-            raise weaponsException404(name_obj='product', _property=model)
-
-        self.db.delete(product)
-        self.db.commit()
-
-    def delete_product(self, product: Product):
-        self.db.delete(product)
-        self.db.commit()
-
-    def add_new_product(self, product: Product)->int:
-        self.db.add(product)
-        self.db.commit()
-        self.db.refresh(product)
-
-        if not product.id:
-            raise products_not_added_exc
-
-        return product.id
-
-    def add_new_property(self, _property: Property):
-        self.db.add(_property)
-        # if _property.product_id:
-        #     raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail='failed add property')
-        self.db.commit()
-
-    def update_product(self, model: str, request: ProductSchema):
-        product = self.db.query(Product).filter(Product.model == model).first()
-        if not product:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                                detail=f'product \'{model.upper()}\' is absent in db!')
-
-        self.db.query(Product).filter(Product.model == model).update(request.model_dump(exclude_unset=True))
-        self.db.commit()
-
-    def modify_product(self, model: str, request: ProductSchemaModify):
-        product = self.db.query(Product).filter(Product.model == model).first()
-        if not product:
-            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND,
-                                detail=f'product \'{model.upper()}\' is absent in db!')
-
-        request_update = request.model_dump(exclude_unset=True)
-        for key, value in request_update.items():
-            setattr(product, key, value)
-        self.db.commit()
+    @property
+    def property_repository(self) ->AbcPropertyRepository:
+        return PropertyRepository(self.db)
